+++ packages/formatters/src/index.ts
@@ -1,7 +1,7 @@
 import './JSON';
 import './YAML';
+import './INI';
 import './Dotenv';
-import './TOML';
 
 // todo: restore rest of the formatters
 // todo: improve & optimize current formatters
